Extract general actions to a class.
These actions do not depend on whether they are run from a specific
view, and can be reused elsewhere: undo / redo, delete selection,
navigations, ....
